added workplace class with methods for queries
